Add warning for hex files
Most flashing tools want hex files so this is an easy mistake to make. Conversion with objcopy is trivial so we don't have to add support for hex. However, we should at least warn the user. The README is quite clear about that but I had only looked at `--help`.

It seems that I'm not the only who made this mistake: https://github.com/MarsTechHAN/ch552tool/issues/2