Add device entries for CH579 and CH32V103.
My CH32V103 with bootloader 2.60 needs an empty WRITE_CMD_V2
command after the payload has been uploaded:  Without this
the verify will fail immediately.

My CH579 device with bootloader 2.7 will fail with this
empty write.

I'm not sure if this write is needed for bootloader 2.60 only,
or if its needed for all CH32V103 devices.  Err'ing on the latter
for now.